feat : register RegexGenerator to generatorFactory
